fix: fix lint error
