+++ pkg/filter/stream/common/http/request_handler.go
@@ -16,6 +16,28 @@
 
 package http
 
+import (
+	"context"
+	"encoding/json"
+	"errors"
+)
+
+type ContextKeyRequestData struct {
+}
+
 type RequestHandler interface {
 	GetEndpoint(name string) (endpoint Endpoint, ok bool)
+}
+
+func GetRequestData(ctx context.Context) (map[string]interface{}, error) {
+	requestData := ctx.Value(ContextKeyRequestData{})
+	if requestData == nil {
+		return nil, errors.New("invalid request body")
+	}
+	conf := make(map[string]interface{})
+	err := json.Unmarshal(requestData.([]byte), &conf)
+	if err != nil {
+		return nil, err
+	}
+	return conf, nil
 }
